+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 
 - Fix variables like `player.element` referring to a wrong player when access in the client context
+- Fix remote crashed when trying to spawn an effect whose sheet is not yet loaded
 
 ## [0.6.1] 2025-11-19